+++ Source/CesiumRuntime/CesiumRuntime.Build.cs
@@ -19,7 +19,8 @@ public class CesiumRuntime : ModuleRules
 
         PrivateIncludePaths.AddRange(
             new string[] {
-              Path.Combine(GetModuleDirectory("Renderer"), "Private")
+              Path.Combine(GetModuleDirectory("Renderer"), "Private"),
+              Path.Combine(GetModuleDirectory("Renderer"), "Internal")
             }
         );