+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Change Log {#changes}
 
-### ? - ?
+### v2.17.0 - 2025-07-01
 
 This is the last release of Cesium for Unreal that will support Unreal Engine v5.3. Future versions will require Unreal Engine v5.4+.